requirements.txt
streamlit==1.28.1
requests==2.31.0
Pillow==10.0.1
pandas==2.1.1
python-dotenv==1.0.0
base64==1.0.0